KYLIN-5848 update modules

+++ src/spark-project/engine-spark/src/main/scala/org/apache/kylin/engine/spark/job/ResourceDetectBeforeSampling.java
@@ -51,7 +51,7 @@ public class ResourceDetectBeforeSampling extends SparkApplication implements Re
         final TableDesc tableDesc = NTableMetadataManager.getInstance(config, project).getTableDesc(tableName);
         Map<String, String> params = NProjectManager.getInstance(config).getProject(project)
                 .getLegalOverrideKylinProps();
-        long rowCount = Long.parseLong(getParam(NBatchConstants.P_SAMPLING_ROWS));
+        int rowCount = Integer.parseInt(getParam(NBatchConstants.P_SAMPLING_ROWS));
         params.put("sampleRowCount", String.valueOf(rowCount));
         final Dataset<Row> dataset = SourceFactory
                 .createEngineAdapter(tableDesc, NSparkCubingEngine.NSparkCubingSource.class)
